Amended README & applied fixes to copier engine.

+++ copier.py
@@ -1,8 +1,8 @@
-from threading import Thread,Lock
+from threading import Thread
 from queue import Queue
-import os,shutil,sqlite3
+#from hashlib import md5,sha256,sha512
+import os,shutil,sqlite3,hashlib
 
-#@dataclass
 class Task:
     """
     An encapsulation of a `Task` that a `CopierThread` will need to process from the queue filled within the `CopierManager` class.
@@ -18,6 +18,41 @@ class Task:
     #END def __init__
 #END class Task
 
+class Hasher:
+    """
+    A basic hash encapsulation that unifies multiple hashing algorithms into one syntax pattern.
+
+    If the hasher is set to 'none', then it simply checks to see if the target exists.
+    """
+    __algorithmDict = {
+        "none":None,
+        "siphash":hash,
+        "md5":hashlib.md5,
+        "sha256":hashlib.sha256,
+        "sha512":hashlib.sha512
+    }
+    __algorithm : str
+    def __init__(self,algorithm:str):
+        if algorithm.lower() in self.__algorithmDict.keys():
+            self.__algorithm=algorithm.lower()
+        else:
+            raise KeyError(f"Key {algorithm} not found.")
+    #END def __init__
+
+    def chceksum(self,file:os.DirEntry) -> str:
+        if self.__algorithmDict[self.__algorithm] is None:
+            return str(os.path.exists(file))
+        else:
+            with open(file,"rb") as f:
+                checksum=self.__algorithmDict[self.__algorithm](f.read())
+            if type(checksum) is int:
+                return str(checksum)
+            else:
+                return checksum.hexdigest()
+    #END def checksum
+
+#END class Hasher
+
 class CopierThread(Thread):
     """
     An individual thread for the Copier process.  Pulls a `Task` from the queue to process, applying the provided options flags as necessary.
@@ -26,14 +61,14 @@ class CopierThread(Thread):
     __queuePointer : Queue[Task]
     __dbConnection : sqlite3.Connection
     __currentTask : Task
-    __lockPointer : Lock
+    __HasherPointer : Hasher
     __options : dict
 
-    def __init__(self,queuePointer:Queue[Task],lockPointer:Lock,options:dict) -> None:
+    def __init__(self,queuePointer:Queue[Task],HasherPointer:Hasher,options:dict) -> None:
         super().__init__()
         self.__queuePointer=queuePointer
         self.__dbConnection=None
-        self.__lockPointer=lockPointer
+        self.__HasherPointer=HasherPointer
         self.__options=options
     #END def __init__
 
@@ -45,6 +80,8 @@ class CopierThread(Thread):
             destFile:os.DirEntry
             destFile=os.path.normpath(f"{self.__currentTask.destination}/{self.__currentTask.source.name}")
 
+            #Get hash of source file.
+
             #TODO:
             #Check if file confliction exists.  If so, apply the appropriate response.
             logStatement="Unknown Error."
@@ -52,16 +89,19 @@ class CopierThread(Thread):
             logAction=self.__currentTask.jobType
             actionStatus="Succeeded"
             logSource=os.path.abspath(self.__currentTask.source)
-            logDest=os.path.abspath(self.__currentTask.destination)
+            logDest=os.path.abspath(destFile)
             execute=True
+            hashSource=self.__HasherPointer.chceksum(self.__currentTask.source)
+            hashDest="None"
             if os.path.exists(destFile):
+                hashDest=self.__HasherPointer.chceksum(destFile)
                 conflictionStatus="Confliction encountered: "
                 collisionMode=self.__options.get("fileConflictMode",0)
-                if collisionMode==0: #Do not process the file if set to 'Do nothing.'
+                if collisionMode==0 and hashSource==hashDest: #Do not process the file if set to 'Do nothing.'
                     execute=False
-                elif collisionMode==1: #Overwrite the file existing in the destination.
-                    pass
-                elif collisionMode==2: #Only process the file if the source is newer than the destination.
+                elif collisionMode==1 and hashSource==hashDest: #Overwrite the file existing in the destination.
+                    execute=False
+                elif collisionMode==2 and hashSource==hashDest: #Only process the file if the source is newer than the destination.
                     if self.__currentTask.source.stat().st_mtime>=destFile.stat().st_mtime:
                         execute=False
             else:
@@ -73,13 +113,21 @@ class CopierThread(Thread):
                     #If mode is move, copy the item, then add a new task to the queue to delete the source file.
                     try:
                         shutil.copy2(self.__currentTask.source,self.__currentTask.destination)
-                        logStatement="Copy success."
-                        if self.__currentTask.jobType=="Move":
-                            self.__currentTask.destination=destFile
-                            self.__currentTask.jobType="Delete"
-                            self.__queuePointer.put(self.__currentTask)
-                            logStatement+="  Source marked for deletion."
-                        #Write to the log that the file was copied.  If it was marked for deletion, then include that statement.
+                        #Get hash of copied file.  If the hash does not match the source, retry the copy.
+                        hashDest=self.__HasherPointer.chceksum(destFile)
+                        if hashSource==hashDest:
+                            logDest=os.path.abspath(destFile)
+                            actionStatus="Success"
+                            logStatement="Copy success."
+                            if self.__currentTask.jobType=="Move":
+                                self.__currentTask.destination=destFile
+                                self.__currentTask.jobType="Delete"
+                                self.__queuePointer.put(self.__currentTask)
+                                logStatement+="  Source marked for deletion."
+                        else:
+                            actionStatus="Failure"
+                            logStatement="Copy failed (checksum mismatch)."
+
                     except OSError: #Currently, if an IO error occurs, skip this file.
                         actionStatus="Failure"
                         logStatement="Copy failed (check privileges/connection)!  Process skipped."
@@ -87,21 +135,32 @@ class CopierThread(Thread):
                         #Retry a set number of times.  If the task cannot be completed, log the error and move on.
                 elif self.__currentTask.jobType=="Delete":
                     #If mode is delete, delete the source file.
-                    try:
-                        os.remove(self.__currentTask.source)
-                        actionStatus="Success"
-                        logStatement="Deletion succeeded."
-                    except OSError:
-                        actionStatus="Failure"
-                        logStatement="Deletion failed (check privileges)!  Process skipped."
-                    #Log the result of the operation to the log DB.
+                    #Validate that the destination file exists, then do a checksum.  If checksum passes, delete; otherwise, enter as new 'Move' task.
+                    hashDest=self.__HasherPointer.chceksum(self.__currentTask.destination)
 
+                    if hashSource==hashDest:
+                        try:
+                            os.remove(self.__currentTask.source)
+                            actionStatus="Success"
+                            logStatement="Deletion succeeded."
+                        except OSError:
+                            actionStatus="Failure"
+                            logStatement="Deletion failed (check privileges)!  Process skipped."
+                        #Log the result of the operation to the log DB.
+                    else:
+                        actionStatus="Failure"
+                        logStatement="Checksum mismatch; retrying move operation."
+                        self.__currentTask.destination=os.path.dirname(self.__currentTask.destination)
+                else:
+                    actionStatus="Failure."
+                    logStatement="Unknown job operation.  Process skipped."
             else:
-                logStatement="Process skipped."
+                actionStatus="Success"
+                logStatement="Skip condition met.  Process skipped."
             self.__queuePointer.task_done()
             #Submit log to the DB.
             with self.__dbConnection:
-                query=f"insert into Completed(action,status,description,source,destination) values ('{logAction}','{actionStatus}','{conflictionStatus}{logStatement}','{logSource}','{logDest}')"
+                query=f"insert into Completed(action,status,description,source,src_checksum,destination,dest_checksum) values ('{logAction}','{actionStatus}','{conflictionStatus}{logStatement}','{logSource}','{hashSource}','{logDest}','{hashDest}')"
                 self.__dbConnection.execute(query)
         self.__dbConnection.close()
     #END def run
@@ -134,15 +193,15 @@ class CopierManager:
     """
     __queue : Queue[Task]
     __threads : list[CopierThread]
-    __Lock : Lock
+    __Hasher : Hasher
     __options : dict
 
     def __init__(self,options:dict):
         self.__options=options
         self.__queue=Queue(maxsize=options.get("queue-maximum",250000))
-        self.__Lock=Lock()
+        self.__Hasher=Hasher(options.get("logType","None"))
 
-        self.__threads=[CopierThread(self.__queue,self.__Lock,options) for thread in range(options.get("threads",1))]
+        self.__threads=[CopierThread(self.__queue,self.__Hasher,options) for thread in range(options.get("threads",1))]
     #END def __init__
 
     def startJob(self,sources:list[str],destination:str):
@@ -155,14 +214,16 @@ class CopierManager:
         db=sqlite3.connect(logDBPath)
         with db:
             db.cursor().execute(
-                "create table if not exists Completed (action varchar(8) not null,status varchar(8) not null,description varchar(256) not null,source varchar(4098) not null,destination varchar(4098) not null);"
+                "create table if not exists Completed (action varchar(8) not null,status varchar(8) not null,description varchar(256) not null,source varchar(4098) not null,src_checksum varchar(512),destination varchar(4098) not null,dest_checksum varchar(512));"
             )
             #Table name: Completed,
             #action : varchar(8) 
             #status : varchar(8)
             #description : varchar(256)
             #source : varchar(4098)
+            #src_checksum : varchar(512)
             #destination : varchar(4098)
+            #dest_checksum : varchar(512)
 
         for thread in self.__threads:
             thread.setDB(logDBPath)
@@ -170,7 +231,9 @@ class CopierManager:
             thread.start()
 
         for source in sources:
-            for task in self.__listDirs(os.path.abspath(source),os.path.abspath(destination)):
+            sourcePath=os.path.abspath(source)
+            destinationPath=os.path.abspath(f"{destination}/{os.path.basename(sourcePath)}")
+            for task in self.__listDirs(sourcePath,destinationPath):
                 self.__queue.put(task)
 
         self.__queue.join()
@@ -236,7 +299,7 @@ def main():
     options={"cli":True}
     with open("./versionData/defaultOptions.json","r") as file:
         options=json.load(file)
-    #defaultOptions=options.copy()
+    defaultOptions=options.copy()
 
     #noShortenedFlags={
     #    "--log":"log"
@@ -249,7 +312,8 @@ def main():
         "--log-destination":"logDest",
         "--hash-algorithm":"logType",
         "--thread-count":"threads",
-        "--conflict":"fileConflictMode"
+        "--conflict":"fileConflictMode",
+        "--retry":"retry"
     }
 
     shortenedFlags={f"-{flag[2]}":flag for flag in flags}
@@ -279,6 +343,10 @@ def main():
             options["log"]=True
             if param=="":
                 param=options["destination"]
+        elif actualFlag=="threads" or actualFlag=="retry" or actualFlag=="fileConflictMode":
+            param=int(param)
+            if param <= 0 and (actualFlag=="threads" or actualFlag=="retry"):
+                param=defaultOptions[actualFlag]
 
         options[actualFlag]=param
     #print(options)
